Update RemoveEdge.ps1
- Script now requires admin privileges to run to prevent errors
- The script will exit if Microsoft Edge is not found on the system
- Updated comments to be more accurate
- Edge now won't be removed from start and desktop because I never got it to work
- Script will now tell you what files are being deleted

+++ RemoveEdge.ps1
@@ -1,10 +1,18 @@
-# Force close Edge
+#Requires -RunAsAdministrator
+
+# Check if Edge is installed
+if (!(Test-Path "${env:ProgramFiles(x86)}\Microsoft\Edge")){
+    Write-Host "Error: Can't Find Microsoft Edge"
+    pause; exit
+}
+
+# Force stop Edge
 Get-Process msedge -ErrorAction SilentlyContinue | Stop-Process
 Start-Sleep 2
 
-# Remove Edge from ProgramFiles, AppData, ProgramData, Start Menu, & Desktop
-Get-Item "${env:ProgramFiles(x86)}\Microsoft\Edge\","$env:LOCALAPPDATA\Packages\Microsoft.MicrosoftEdge*","$env:ProgramData\Microsoft\Windows\Start Menu\Programs\Microsoft Edge*","$HOME\Desktop\Microsoft Edge*" | % {
-    Remove-Item "$_" -Recurse -Force
+# Remove Edge from ProgramFiles & AppData
+Get-Item "${env:ProgramFiles(x86)}\Microsoft\Edge\","$env:LOCALAPPDATA\Packages\Microsoft.MicrosoftEdge*" | % {
+Remove-Item "$_" -Recurse -Force -Verbose
 }
 
 # Create registry file that tells Windows you have the old non-Chromium Edge browser (disables updates)
